# Speed up concretization