This isn't a rework of the concretizer but it speeds things up a LOT.
    
The main culprits were:
  1. Variant code and `provider_index` were calling `spec.package` when they
     could use `spec.package_class`
      - `spec.package` looks up a package instance by `Spec`, which requires a
        (fast-ish but not that fast) DAG compare.
      - `spec.package_class` just looks up the package's class by name, and you
          should use this when all you need is metadata (most of the time).
      - not really clear that the current way packages are looked up is
        necessary -- we can consider refactoring that in the future.
    
  2. `Repository.repo_for_pkg` parses a `str` argument into a `Spec` when
     called with one, via `@_autospec`, but this is not needed.
     - Add some faster code to handle strings directly and avoid parsing
   
This **speeds up concretization ~~3-6x~~  3-9x** in my limited tests.  Probably more for bigger DAGs.  Still not super fast but much more bearable:

Before:
  - `spack spec xsdk` took 33.6s
  - `spack spec dealii` took 1m39s
  - `spack spec r-rminer` takes 😱 

After:
  - `spack spec xsdk` takes ~~11.4s~~ 6.8s
  - `spack spec dealii` takes ~~18.6s~~ 10.8s
  - `spack spec r-rminer` takes 1m42s (this one is kind of an upper bound for us so far)

@adamjstewart @davydden 😄 ☝️