@alalazo 2017-10-12
Restarted CI (python 3.6 errored)