@davydden 2017-10-12
> spack spec dealii took 1m39s
vs
> spack spec dealii takes 18.6s

Awesome! 😄 🎉