@tgamblin 2017-10-12
ok stomped out some more needless lookups/parsings in `provider_index` and in `Repository`.  Got it down to 10s for dealii and 6.8 for xsdk.