@healther 2017-10-12
Found some problem concretising our meta-package, not 100% sure if this should go here, but I know that it worked with the concretiser from the end of August

```
Concretized
--------------------------------
==> Error: An unsatisfiable variant constraint has been detected for spec:

    pcre@8.40%clang@9.0.0-apple~jit+utf arch=darwin-macos-x86_64 


while trying to concretize the partial spec:

    git@2.14.1%clang@9.0.0-apple arch=darwin-macos-x86_64 
        ^curl@7.56.0%clang@9.0.0-apple arch=darwin-macos-x86_64 
            ^nghttp2@1.26.0%clang@9.0.0-apple arch=darwin-macos-x86_64 
            ^openssl@1.0.2k%clang@9.0.0-apple arch=darwin-macos-x86_64 
                ^zlib@1.2.11%clang@9.0.0-apple+pic+shared arch=darwin-macos-x86_64 
        ^gettext@0.19.8.1%clang@9.0.0-apple+bzip2+curses+git~libunistring+libxml2+tar+xz arch=darwin-macos-x86_64 
            ^bzip2@1.0.6%clang@9.0.0-apple+shared arch=darwin-macos-x86_64 
            ^libxml2@2.9.4%clang@9.0.0-apple~python arch=darwin-macos-x86_64 
                ^pkg-config@0.29.2%clang@9.0.0-apple+internal_glib arch=darwin-macos-x86_64 
                ^xz@5.2.3%clang@9.0.0-apple arch=darwin-macos-x86_64 
            ^ncurses@6.0%clang@9.0.0-apple~symlinks arch=darwin-macos-x86_64 
            ^tar
        ^m4@1.4.18%clang@9.0.0-apple+sigsegv arch=darwin-macos-x86_64 
            ^libsigsegv


git requires pcre variant +jit, but spec asked for ~jit
```

First of all +2 for the nicer error output, and I think it originates with the way the dependency is introduced in ```git/package.py```:
```
153:     depends_on('pcre', when='@:2.13')
154:     depends_on('pcre+jit', when='@2.14:')
```